+++ packages/apollos-ui-connected/src/CampaignItemListFeature/index.js
@@ -11,7 +11,9 @@ import {
   styled,
   withIsLoading,
 } from '@apollosproject/ui-kit';
-import { LiveConsumer } from '..';
+
+import ListItemCard from '../ListItemCard';
+import { LiveConsumer } from '../live';
 
 const Title = styled(
   ({ theme }) => ({
@@ -31,8 +33,9 @@ const ListItemComponent = ({ contentId, ...item }) => (
   <LiveConsumer contentId={contentId}>
     {(liveStream) => {
       const isLive = !!(liveStream && liveStream.isLive);
-      const labelText = isLive ? 'Live' : item.labelText;
-      return <FeaturedCard isLive={isLive} {...item} labelText={labelText} />;
+      return (
+        <ListItemCard Component={FeaturedCard} isLive={isLive} {...item} />
+      );
     }}
   </LiveConsumer>
 );
